backend/requirements.txt
google-generativeai
flask
flask-cors
nltk
numpy
tensorflow-cpu
pickle-mixin
gunicorn
